WEB/Chall/SiakGG-Pekan Ristek/siakgg/src/requirements.txt
flask
psycopg2-binary
pyjwt
bcrypt
